requirements.txt
Django==1.11.7
django-crispy-forms==1.7.0
django-solo==1.1.2
pytz==2017.3
